Add **kwargs to spreadsheet.delete()
Without passing the `supportsTeamDrives=True` flag to spreadsheet.delete() on a spreadsheet located in a Team Drive, a `googleapiclient.errors.HttpError` will be raised claiming the file can't be found. I worked around this by using:

```
gc.drive.delete(spreadsheet.id, supportsTeamDrives=True)
```

But it would be nice to use this on the spreadsheet object like much of the other API wrapper functions.